requirements.txt
flask>=2.0.0
line-bot-sdk==3.16.0
google-generativeai
python-dotenv
